+++ src/main/java/uk/co/real_logic/agrona/concurrent/ringbuffer/ManyToOneRingBuffer.java
@@ -86,7 +86,7 @@ public class ManyToOneRingBuffer implements RingBuffer
     /**
      * {@inheritDoc}
      */
-    public boolean write(final int msgTypeId, final DirectBuffer srcBuffer, final int srcIndex, final int length)
+    public boolean write(final int msgTypeId, final DirectBuffer srcBuffer, final long srcIndex, final long length)
     {
         checkTypeId(msgTypeId);
         checkMsgLength(length);
@@ -316,7 +316,7 @@ public class ManyToOneRingBuffer implements RingBuffer
         return allZeros;
     }
 
-    private void checkMsgLength(final int length)
+    private void checkMsgLength(final long length)
     {
         if (length > maxMsgLength)
         {
